Editor handles remote changes properly
The earlier implementation was a hack. The editor was only able to
handle changes from the server that were simple inserts or simple
deletions. Any change that involved replacing text would not update
properly. This is now fixed so that all character changes correctly
update.

The editor also ensures that the text ends with a newline. This is so
that any changes made from the iOS demo stay compatible with the web
demo (the quill editor requires that the text end with a newline).